setting of the lifetime of the generated images has been moved to the file configuration.properties; bug fixed
git-svn-id: http://meta-editor.googlecode.com/svn/trunk@495 50e9bae0-2ccb-1508-233a-4bfc89a952fd

+++ MEditor/src/cz/fi/muni/xkremser/editor/server/handler/ScanFolderHandler.java
@@ -184,13 +184,27 @@ public class ScanFolderHandler
      * @throws ActionException
      */
     private void removeOldImages() throws ActionException {
-        try {
-            ArrayList<String> oldImages = imageResolverDAO.cacheAgeingProcess();
-            for (String oldImage : oldImages) {
-                new File(oldImage).delete();
+
+        int numberOfDays = Integer.MIN_VALUE;
+        String numberString = configuration.getGenImagesLifetime();
+
+        if (numberString != null) {
+            try {
+                numberOfDays = Integer.parseInt(numberString);
+                if (numberOfDays > 0) {
+                    try {
+
+                        ArrayList<String> oldImages = imageResolverDAO.cacheAgeingProcess(numberOfDays);
+                        for (String oldImage : oldImages) {
+                            new File(oldImage).delete();
+                        }
+
+                    } catch (DatabaseException e) {
+                        throw new ActionException(e);
+                    }
+                }
+            } catch (NumberFormatException nfe) {
             }
-        } catch (DatabaseException e) {
-            throw new ActionException(e);
         }
     }
